propagation v4
i think i'm getting it, at last…

this restores the appraoch of assigning values immediately

there is no substantial change (ignoring arugments passed, etc) aside from moving the update from dequeue to enqueue

this works, roughly, as when updating it's, roughly, a search to find some other watch for that literal
the only time anything happens is when the search fails, which means whatever the other watch is much follow, whether none or conflicting
this does not rely on what they current watches point to, really

most simply, the other watch is ignored, and if there's no other candidate for the current watch then whatever the other watch is much be

etc etc …

things to tidy, and may make this a config option, as while i prefer the config version it is a choice

+++ src/otter_lib/config/mod.rs
@@ -1,8 +1,8 @@
 pub mod defaults;
 
 pub type ActivityType = f64;
-pub type DecayFrequency = usize;
-pub type GlueStrength = u32;
+pub type DecayFrequency = u8;
+pub type GlueStrength = u8;
 pub type LubyConstant = usize;
 pub type PolarityLean = f64;
 pub type RandomChoiceFrequency = f64;
